test: move s3 mock to root setup
