+++ packages/blob-storage-manager/setup.ts
@@ -1,4 +1,3 @@
-import { Readable } from "stream";
 import { afterAll, vi } from "vitest";
 
 import { SWARM_REFERENCE } from "./test/fixtures";
@@ -69,61 +68,6 @@ vi.mock("@ethersphere/bee-js", async (importOriginal) => {
   };
 });
 
-vi.mock("@aws-sdk/client-s3", async (originalImport) => {
-  // eslint-disable-next-line @typescript-eslint/consistent-type-imports
-  const original = await originalImport<typeof import("@aws-sdk/client-s3")>();
-  const storage: Record<string, Uint8Array> = {};
-
-  return {
-    S3Client: vi.fn().mockImplementation((_) => {
-      return {
-        send(
-          c:
-            | typeof original.GetObjectCommand
-            | typeof original.PutObjectCommand
-            | typeof original.DeleteObjectCommand
-            | typeof original.HeadBucketCommand
-        ) {
-          if (c instanceof original.HeadBucketCommand) {
-            if (c.input.Bucket !== "blobscan-s3-bucket") {
-              throw new Error("No such bucket");
-            }
-
-            return;
-          }
-
-          if (c instanceof original.GetObjectCommand) {
-            const binaryData = storage[c.input.Key as string];
-
-            if (!binaryData) {
-              throw new Error("no such key exists");
-            }
-            return {
-              Body: binaryData ? Readable.from(binaryData) : undefined,
-            };
-          }
-
-          if (c instanceof original.PutObjectCommand) {
-            if (Buffer.isBuffer(c.input.Body)) {
-              const binaryData = new Uint8Array(c.input.Body);
-              storage[c.input.Key as string] = binaryData;
-            }
-          }
-
-          if (c instanceof original.DeleteObjectCommand) {
-            if (storage[c.input.Key as string]) {
-              delete storage[c.input.Key as string];
-            }
-          }
-        },
-      };
-    }),
-    GetObjectCommand: original.GetObjectCommand,
-    PutObjectCommand: original.PutObjectCommand,
-    DeleteObjectCommand: original.DeleteObjectCommand,
-    HeadBucketCommand: original.HeadBucketCommand,
-  };
-});
 afterAll(() => {
   vi.clearAllMocks();
 });
